move register scheduler to schedule
