+++ server/config/config_test.go
@@ -53,8 +53,6 @@ func TestBadFormatJoinAddr(t *testing.T) {
 
 func TestReloadConfig(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
-	RegisterScheduler("shuffle-leader")
 	opt, err := newTestScheduleOption()
 	re.NoError(err)
 	storage := storage.NewStorageWithMemoryBackend()
@@ -64,22 +62,10 @@ func TestReloadConfig(t *testing.T) {
 	opt.GetPDServerConfig().UseRegionStorage = true
 	re.NoError(opt.Persist(storage))
 
-	// Add a new default enable scheduler "shuffle-leader"
-	DefaultSchedulers = append(DefaultSchedulers, SchedulerConfig{Type: "shuffle-leader"})
-	defer func() {
-		DefaultSchedulers = DefaultSchedulers[:len(DefaultSchedulers)-1]
-	}()
-
 	newOpt, err := newTestScheduleOption()
 	re.NoError(err)
 	re.NoError(newOpt.Reload(storage))
-	schedulers := newOpt.GetSchedulers()
-	re.Len(schedulers, len(DefaultSchedulers))
-	re.True(newOpt.IsUseRegionStorage())
-	for i, s := range schedulers {
-		re.Equal(DefaultSchedulers[i].Type, s.Type)
-		re.False(s.Disable)
-	}
+
 	re.Equal(5, newOpt.GetMaxReplicas())
 	re.Equal(uint64(10), newOpt.GetMaxSnapshotCount())
 	re.Equal(int64(512), newOpt.GetMaxMovableHotPeerSize())
@@ -87,7 +73,6 @@ func TestReloadConfig(t *testing.T) {
 
 func TestReloadUpgrade(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	opt, err := newTestScheduleOption()
 	re.NoError(err)
 
@@ -111,7 +96,6 @@ func TestReloadUpgrade(t *testing.T) {
 
 func TestReloadUpgrade2(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	opt, err := newTestScheduleOption()
 	re.NoError(err)
 
@@ -133,7 +117,6 @@ func TestReloadUpgrade2(t *testing.T) {
 
 func TestValidation(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	cfg := NewConfig()
 	re.NoError(cfg.Adjust(nil, false))
 
@@ -163,8 +146,6 @@ func TestValidation(t *testing.T) {
 
 func TestAdjust(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
-	RegisterScheduler("random-merge")
 	cfgData := `
 name = ""
 lease = 0
@@ -226,32 +207,6 @@ max-merge-region-keys = 400000
 	re.NoError(err)
 	re.Contains(cfg.WarningMsgs[0], "Config contains undefined item")
 	re.Equal(40*10000, int(cfg.Schedule.GetMaxMergeRegionKeys()))
-	// Check misspelled schedulers name
-	cfgData = `
-name = ""
-lease = 0
-
-[[schedule.schedulers]]
-type = "random-merge-schedulers"
-`
-	cfg = NewConfig()
-	meta, err = toml.Decode(cfgData, &cfg)
-	re.NoError(err)
-	err = cfg.Adjust(&meta, false)
-	re.Error(err)
-	// Check correct schedulers name
-	cfgData = `
-name = ""
-lease = 0
-
-[[schedule.schedulers]]
-type = "random-merge"
-`
-	cfg = NewConfig()
-	meta, err = toml.Decode(cfgData, &cfg)
-	re.NoError(err)
-	err = cfg.Adjust(&meta, false)
-	re.NoError(err)
 
 	cfgData = `
 [metric]
@@ -293,7 +248,6 @@ tso-update-physical-interval = "15s"
 
 func TestMigrateFlags(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	load := func(s string) (*Config, error) {
 		cfg := NewConfig()
 		meta, err := toml.Decode(s, &cfg)
@@ -331,7 +285,6 @@ disable-make-up-replica = false
 
 func TestPDServerConfig(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	tests := []struct {
 		cfgData          string
 		hasErr           bool
@@ -398,7 +351,6 @@ dashboard-address = "foo"
 
 func TestDashboardConfig(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	cfgData := `
 [dashboard]
 tidb-cacert-path = "/path/ca.pem"
@@ -438,7 +390,6 @@ tidb-cert-path = "/path/client.pem"
 
 func TestReplicationMode(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	cfgData := `
 [replication-mode]
 replication-mode = "dr-auto-sync"
@@ -474,7 +425,6 @@ wait-store-timeout = "120s"
 
 func TestHotHistoryRegionConfig(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	cfgData := `
 [schedule]
 hot-regions-reserved-days= 30
@@ -497,7 +447,6 @@ hot-regions-write-interval= "30m"
 
 func TestConfigClone(t *testing.T) {
 	re := require.New(t)
-	registerDefaultSchedulers()
 	cfg := &Config{}
 	cfg.Adjust(nil, false)
 	re.Equal(cfg, cfg.Clone())
@@ -528,10 +477,4 @@ func newTestScheduleOption() (*PersistOptions, error) {
 	}
 	opt := NewPersistOptions(cfg)
 	return opt, nil
-}
-
-func registerDefaultSchedulers() {
-	for _, d := range DefaultSchedulers {
-		RegisterScheduler(d.Type)
-	}
 }
